Hw2.Log

Documentation

data MessageType

Constructors

Info

Warning

Error Int

type TimeStamp = Int

data LogMessage

Constructors

LogMessage MessageType TimeStamp String

Unknown String

data MessageTree

Constructors

Leaf

Node MessageTree LogMessage MessageTree

testParse :: (String -> [LogMessage]) -> Int -> FilePath -> IO [LogMessage]

testParse p n f tests the log file parser p by running it on the first n lines of file f.

testWhatWentWrong :: (String -> [LogMessage]) -> ([LogMessage] -> [String]) -> FilePath -> IO [String]

testWhatWentWrong p w f tests the log file parser p and warning message extractor w by running them on the log file f.